refactor(workspace): split multi-root settings boundaries
