+++ crates/agent-ui/src/components/chat/WorkspaceResourceSettingsDrawer.tsx
@@ -1,8 +1,8 @@
-export {
-  type WorkspaceProjectRootAccess,
-  type WorkspaceProjectRootClient,
-  type WorkspaceProjectRootDraft,
-  type WorkspaceProjectRootGrant,
-  type WorkspaceProjectRootState,
-  WorkspaceProjectSettingsModal as WorkspaceResourceSettingsDrawer,
-} from "./WorkspaceProjectSettingsModal";
+export type {
+  WorkspaceProjectRootAccess,
+  WorkspaceProjectRootClient,
+  WorkspaceProjectRootDraft,
+  WorkspaceProjectRootGrant,
+  WorkspaceProjectRootState,
+} from "@liveagent/ui/contracts/workspaceProjectRoots";
+export { WorkspaceProjectSettingsModal as WorkspaceResourceSettingsDrawer } from "./WorkspaceProjectSettingsModal";
